1_Prompt_Chain/requirement.txt
langchain-google-genai
python-dotenv
langchain-core
